backend_publication/requirements.txt
django
djangorestframework
django-filters
requests
